Address the coding style issues
