+++ Code/Framework/AzFramework/AzFramework/Session/SessionRequests.cpp
@@ -8,7 +8,7 @@
 
 #include <AzCore/Serialization/EditContext.h>
 #include <AzCore/Serialization/SerializeContext.h>
-#include <AzFramework/Session/ISessionRequests.h>
+#include <AzFramework/Session/SessionRequests.h>
 #include <AzFramework/Session/SessionConfig.h>
 
 namespace AzFramework
